organize_code/requirements.txt
av==12.0.0
clip==1.0
cycler==0.12.1
dataclasses_json==0.6.7
decord==0.6.0
e3nn==0.5.6
einops==0.8.1
fire==0.5.0
Flask==3.1.1
flask_cors==6.0.1
gradio==5.34.1
gsplat==0.1.12
hdbscan==0.8.40
huggingface_hub==0.29.3
hydra-core==1.3.2
imageio==2.33.1
imagesize==1.4.1
iopath==0.1.10
jsonargparse==4.24.1
lightning==2.0.9.post0
loguru==0.7.3
matplotlib==3.10.3
mediapy==1.2.4
numpy==2.3.0
omegaconf==2.3.0
open3d==0.18.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
plyfile==1.1.2
point_cloud_utils==0.34.0
pycocotools==2.0.8
pymeshlab==2023.12.post3
pytest==7.2.0
pytorch3d==0.7.8
rich==14.0.0
scene==0.0.1
scikit_learn==1.3.2
scipy==1.15.3
segment_anything==1.0
setuptools==75.8.0
skimage==0.0
splines==0.3.0
strawberry==3.0
submitit==1.5.3
tensordict==0.8.3
tinycudann==1.7
torch==2.5.1+cu121
torchmetrics==1.7.0
torchvision==0.20.1+cu121
tqdm==4.67.1
trimesh==4.1.0
typing_extensions==4.14.0
viser==0.2.3
wandb==0.20.1
